refactor: createPlaylist
