+++ src/App.js
@@ -1,7 +1,7 @@
 import "./app.css";
-import Header from "./components/header/Header";
-import Sidebar from "./components/sidebar/Sidebar";
-import CreatePlaylist from "./pages/CreatePlaylist";
+import Header from "./components/header";
+import Sidebar from "./components/sidebar";
+import CreatePlaylist from "./pages/CreatePlaylist/CreatePlaylist";
 
 function App() {
   return (
