--- src/components/button/Button.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Button(props, ...rest) {
-  return <button {...rest}>{props}</button>;
-}
