+++ src/components/button/index.js
@@ -1,0 +1,6 @@
+import React from "react";
+import "./button.css";
+
+export default function Button({ name }) {
+  return <button>{name}</button>;
+}
