+++ src/components/image/image.css
@@ -1,0 +1,6 @@
+img {
+  width: 180px;
+  height: 180px;
+  box-shadow: inset 0 -3em 3em rgba(0, 0, 0, 0.1), 0 0 0 0 rgb(255, 255, 255),
+    0.3em 0.3em 1em rgba(0, 0, 0, 0.3);
+}
