+++ src/components/image/index.js
@@ -1,0 +1,6 @@
+import React from "react";
+import "./image.css";
+
+export default function Image(props) {
+  return <img src={props.src} alt={props.alt} />;
+}
