+++ src/components/tagP/index.js
@@ -1,0 +1,6 @@
+import React from "react";
+import "./tagP.css";
+
+export default function TagP(props) {
+  return <p className={props.style}>{props.value}</p>;
+}
