+++ src/components/tagP/tagP.css
@@ -1,0 +1,7 @@
+p {
+  margin: 10px 0;
+}
+
+.details-title {
+  font-size: 30px;
+}
