--- src/pages/createPlaylist.css
@@ -1,51 +1,0 @@
-.create-playlist {
-  background-color: #eeee;
-  width: 100%;
-}
-
-.create-playlist > h1 {
-  padding: 30px;
-}
-
-.create-playlist-detail {
-  display: flex;
-  padding-left: 30px;
-}
-
-.create-playlist-detail > img {
-  width: 180px;
-  height: 180px;
-  box-shadow: inset 0 -3em 3em rgba(0, 0, 0, 0.1), 0 0 0 0 rgb(255, 255, 255),
-    0.3em 0.3em 1em rgba(0, 0, 0, 0.3);
-}
-
-.create-playlist-detail > .details {
-  margin-left: 30px;
-}
-
-.create-playlist-detail > .details > p {
-  margin: 10px 0;
-}
-
-.create-playlist-detail > .details > .details-title {
-  font-size: 30px;
-}
-
-.create-playlist-detail > .details > button {
-  margin-top: 10px;
-  border-style: none;
-  padding: 8px;
-  background-color: #c4c4c4;
-  font-weight: bold;
-  width: 120px;
-  border-radius: 6px;
-  letter-spacing: 1.5px;
-  cursor: pointer;
-}
-
-.create-playlist-detail > .details > button:hover {
-  background-color: grey;
-  box-shadow: inset 0 -3em 3em rgba(0, 0, 0, 0.1), 0 0 0 0 rgb(255, 255, 255),
-    0.3em 0.3em 1em rgba(0, 0, 0, 0.3);
-  color: whitesmoke;
-}
